Fixes based on PR reviews
The removal of the `test` script looks dodgy, but the test for the
export conversion are done in the node module. Since this service
just wraps the module, I don't feel that there's much to test in this
service.
I had originally put in the `test` script as a placeholder.